trainer/requirements.txt
pandas
numpy
scikit-learn
sqlalchemy
pymysql
python-dotenv